[6868] Extract the project image part of the domain into a dedicated module
Bug: https://github.com/eclipse-sirius/sirius-web/issues/6868
