+++ CHANGELOG.adoc
@@ -70,6 +70,8 @@ Imports of these classes must be updated to:
 * `org.eclipse.sirius.web.images.domain.services.api.IImageCreationService`
 * `org.eclipse.sirius.web.images.domain.events.ImageCreatedEvent`
 - https://github.com/eclipse-sirius/sirius-web/issues/6194[#6194] [core] Remove the field `ErrorPayload#message` in favor of `ErrorPayload#messages`.
+- https://github.com/eclipse-sirius/sirius-web/issues/6868[#6868] [sirius-web] The project-image domain has moved to the dedicated Maven module `sirius-web-projects-images-domain`.
+Downstream consumers must declare this dependency and replace the package prefix `org.eclipse.sirius.web.domain.boundedcontexts.projectimage` with `org.eclipse.sirius.web.projects.images.domain`, including its `event`, `repositories`, `services`, and `services.api` subpackages.
 
 
 === Dependency update
