+++ packages/projects/backend/sirius-web-projects-images-domain/.settings/org.eclipse.core.runtime.prefs
@@ -1,0 +1,4 @@
+#Mon Sep 24 15:04:19 CEST 2007
+eclipse.preferences.version=1
+line.separator=\n
+
