+++ packages/projects/backend/sirius-web-projects-images-domain/.settings/org.eclipse.jdt.apt.core.prefs
@@ -1,0 +1,3 @@
+eclipse.preferences.version=1
+org.eclipse.jdt.apt.aptEnabled=false
+
