+++ packages/projects/backend/sirius-web-projects-images-domain/.settings/org.eclipse.m2e.core.prefs
@@ -1,0 +1,5 @@
+activeProfiles=pom.xml
+eclipse.preferences.version=1
+resolveWorkspaceProjects=true
+version=1
+
