+++ packages/projects/backend/sirius-web-projects-images-domain/.settings/org.springframework.ide.eclipse.boot.prefs
@@ -1,0 +1,3 @@
+eclipse.preferences.version=1
+spring.boot.launch.profile.history=dev;
+
