+++ packages/projects/backend/sirius-web-projects-images-domain/src/test/resources/logback-test.xml
@@ -1,0 +1,4 @@
+<configuration>
+  <logger name="com.tngtech.archunit" level="INFO" />
+</configuration>
+
