+++ packages/releng/backend/sirius-components-test-coverage/pom.xml
@@ -453,6 +453,11 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.sirius</groupId>
+			<artifactId>sirius-web-projects-images-domain</artifactId>
+			<version>2026.7.3</version>
+		</dependency>
+		<dependency>
+			<groupId>org.eclipse.sirius</groupId>
 			<artifactId>sirius-web-application</artifactId>
 			<version>2026.7.3</version>
 		</dependency>
