+++ packages/sirius-web/backend/sirius-web-application/pom.xml
@@ -54,6 +54,11 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.sirius</groupId>
+			<artifactId>sirius-web-projects-images-domain</artifactId>
+			<version>2026.7.3</version>
+		</dependency>
+		<dependency>
+			<groupId>org.eclipse.sirius</groupId>
 			<artifactId>sirius-components-graphql-api</artifactId>
 			<version>2026.7.3</version>
 		</dependency>
