--- packages/sirius-web/backend/sirius-web-domain/src/main/java/org/eclipse/sirius/web/domain/boundedcontexts/projectimage/event/IProjectImageEvent.java
@@ -1,23 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2024 Obeo.
- * This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v2.0
- * which accompanies this distribution, and is available at
- * https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     Obeo - initial API and implementation
- *******************************************************************************/
-package org.eclipse.sirius.web.domain.boundedcontexts.projectimage.event;
-
-import org.eclipse.sirius.web.domain.events.IDomainEvent;
-
-/**
- * Interface implemented by all the domain events of the project image bounded context.
- *
- * @author sbegaudeau
- */
-public interface IProjectImageEvent extends IDomainEvent {
-}
