--- packages/sirius-web/backend/sirius-web-domain/src/main/java/org/eclipse/sirius/web/domain/boundedcontexts/projectimage/services/api/IProjectImageDeletionService.java
@@ -1,29 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2024, 2026 Obeo.
- * This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v2.0
- * which accompanies this distribution, and is available at
- * https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     Obeo - initial API and implementation
- *******************************************************************************/
-package org.eclipse.sirius.web.domain.boundedcontexts.projectimage.services.api;
-
-import java.util.UUID;
-
-import org.eclipse.sirius.components.events.ICause;
-import org.eclipse.sirius.web.core.domain.results.IResult;
-
-/**
- * Used to delete project images.
- *
- * @author sbegaudeau
- */
-public interface IProjectImageDeletionService {
-    IResult<Void> deleteProjectImage(ICause cause, UUID projectImageId);
-
-    IResult<Void> deleteAllProjectImages(ICause cause, String projectId);
-}
