--- packages/sirius-web/backend/sirius-web-domain/src/main/java/org/eclipse/sirius/web/domain/boundedcontexts/projectimage/services/api/IProjectImageUpdateService.java
@@ -1,27 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2024, 2026 Obeo.
- * This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v2.0
- * which accompanies this distribution, and is available at
- * https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     Obeo - initial API and implementation
- *******************************************************************************/
-package org.eclipse.sirius.web.domain.boundedcontexts.projectimage.services.api;
-
-import java.util.UUID;
-
-import org.eclipse.sirius.components.events.ICause;
-import org.eclipse.sirius.web.core.domain.results.IResult;
-
-/**
- * Used to update project images.
- *
- * @author sbegaudeau
- */
-public interface IProjectImageUpdateService {
-    IResult<Void> renameProjectImage(ICause cause, UUID projectImageId, String newLabel);
-}
